mails/requirements.txt
certifi==2022.5.18.1
elastic-transport==8.1.2
elasticsearch==8.2.2
python-dotenv==0.20.0
schedule==1.1.0
urllib3==1.26.9
